Add a means of restricting the Key encoding for Map types
For certain AMQP types like MessageAnnotations and ApplicationProperties
the key value of the Map is restricted.  This change adds a mechanism
for the encoder to have the key type provided and only use that meaning
the encode can run faster as the type is known and the encode would fail
if invalid types were placed into the map for encoding.