Create a CompositeReadableBuffer to aggregate Transfer chunks
When a multiple frame Transfer is received we want to aggregate those
chunks into a ReadableBuffer that can be decoded or stored without
recombining the chunks into a single large array.  The codec must be
refactored to accept a ReadableBuffer for decoding such composite
buffers.  Provide means of doing no-copy recv and send using
ReadableBuffer instances to allow more efficient send and receive
operations and reduce GC overhead when the user knows that the bytes are
not reused or shared.